Rename: 내비게이션바 스타일드 컴포넌트 이름 수정
