+++ src/components/header/Logo.jsx
@@ -1,6 +1,5 @@
-import styled from 'styled-components';
 import logoImg from '../../assetes/logoIcon.svg';
-import * as S from './MainBox.style';
+import * as S from './Logo.style';
 
 export default function Logo() {
   return (
