+++ src/components/header/Logo.style.js
@@ -1,12 +1,12 @@
-import * as S from './MainBox.style';
+import styled from 'styled-components';
 
-const LogoStyle = styled.div`
+export const LogoStyle = styled.div`
   display: flex;
   align-items: center;
   gap: 8px;
 `;
 
-const LogoText = styled.div`
+export const LogoText = styled.div`
   color: var(--gray-light-gray-90, #4a494f);
   text-align: center;
   font-size: 19.971px;
@@ -14,7 +14,7 @@ const LogoText = styled.div`
   line-height: normal;
 `;
 
-const LogoImg = styled.div`
+export const LogoImg = styled.div`
   width: 28px;
   height: 28px;
   fill: var(--primary-purple-50, #a64eff);
